+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+## v1.0.94（2020-06-30）
+
+### Bug fixes
+
+- dy-form 修复编辑情况部分值丢失的问题
+
+## v1.0.93（2020-06-17）
+
+### Features
+
+- dy-form 增加 validate-failed 事件
+
+## v1.0.92（2020-06-19）
+
+### Bug fixes
+
+- AdvancedTable 的自定表头优化
+
+## v1.0.90（2020-06-17）
+
+### Features
+
+- AdvancedTable page-sizes 的属性增加
+
+## v1.0.87（2020-06-16）
+
+### Bug fixes
+
+- 面包屑导航文字的鼠标样式
+
 ## v1.0.87（2020-06-16）
 
 ### Bug fixes